+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.8.1](https://github.com/carlonicora/nextjs-jsonapi/compare/v1.8.0...v1.8.1) (2025-12-21)
+
+### 🐛 Bug Fixes
+
+* standardize module name casing in i18n keys and form schema generation ([680179d](https://github.com/carlonicora/nextjs-jsonapi/commit/680179d9f232f8a57db033e3fc021c9eb2bdb10f))
+
+## [1.8.0](https://github.com/carlonicora/nextjs-jsonapi/compare/v1.7.6...v1.8.0) (2025-12-21)
+
+### 🚀 Features
+
+* enhance relationship handling by adding support for edge properties and updating imports ([1aae701](https://github.com/carlonicora/nextjs-jsonapi/commit/1aae701a3300830e161c3cd3465210e12ad00489))
+
 ## [1.7.6](https://github.com/carlonicora/nextjs-jsonapi/compare/v1.7.5...v1.7.6) (2025-12-20)
 
 ### 🐛 Bug Fixes